requirements.txt
APScheduler==3.6.3
asgiref==3.5.0
bcrypt==3.2.0
cachetools==4.2.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
clipboard==0.0.4
colorama==0.4.4
cycler==0.11.0
distlib==0.3.4
Django==4.0.3
dnspython==2.2.1
email-validator==1.1.3
filelock==3.6.0
Flask==2.0.3
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
fonttools==4.32.0
greenlet==1.1.2
heroku==0.1.4
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.1
keyboard==0.13.5
kiwisolver==1.4.2
MarkupSafe==2.1.1
matplotlib==3.5.1
mouse==0.7.1
numpy==1.22.3
packaging==21.3
Pillow==9.0.1
platformdirs==2.5.2
plyer==2.0.0
pycparser==2.21
pygame==2.1.0
pyparsing==3.0.8
pyperclip==1.8.2
pyTelegramBotAPI==4.5.0
python-dateutil==2.8.2
python-telegram-bot==13.11
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.32
sqlparse==0.4.2
termcolor==1.1.0
tornado==6.1
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
virtualenv==20.14.1
Werkzeug==2.0.3
WTForms==3.0.1